+++ 17_PhuongDong_12_HienDuy/frmLoaiPhong.cs
@@ -110,23 +110,36 @@ namespace _17_PhuongDong_12_HienDuy
             {
                 if (flag == 1)
                 {
-                    sql = "insert into LoaiPhong values('" + txtMaPH.Text + "',N'" + txtLoaiPhong.Text + "')";
-                    MessageBox.Show("Bạn có muốn thêm?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    DialogResult kq = MessageBox.Show("Bạn có muốn thêm?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);  
+                   if (kq == DialogResult.Yes)
+                    {
+                        sql = "insert into LoaiPhong values('" + txtMaPH.Text + "',N'" + txtLoaiPhong.Text + "')";
+                    }                
                 }
                 else if (flag == 2)
                 {
-                    sql = "update LoaiPhong set MaPhong = '" + txtMaPH.Text + "', LoaiPhong = N'" + txtLoaiPhong.Text + "' where MaPhong = '" + txtMaPH.Text + "'";
-                    MessageBox.Show("Bạn có muốn sửa phòng vừa chọn?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    DialogResult kq = MessageBox.Show("Bạn có muốn sửa phòng vừa chọn?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                   if (kq == DialogResult.Yes)
+                    {
+                        sql = "update LoaiPhong set MaPhong = '" + txtMaPH.Text + "', LoaiPhong = N'" + txtLoaiPhong.Text + "' where MaPhong = '" + txtMaPH.Text + "'";
+                    }                   
                 }
                 else
                 {
-                    sql = "delete from LoaiPhong where MaPhong = '" + txtMaPH.Text + "'";
-                    MessageBox.Show("Bạn có muốn xóa phòng vừa chọn?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    DialogResult kq = MessageBox.Show("Bạn có muốn xóa phòng vừa chọn?", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                   if (kq == DialogResult.Yes)
+                    {
+                        sql = "delete from LoaiPhong where MaPhong = '" + txtMaPH.Text + "'";
+                    }                    
                 }
+                if (sql != "")
                 if (c.CapNhatDuLieu(sql) != 0)
                 {
-                    MessageBox.Show("Cập nhật thành công!", "Thông báo");
-                    frmLoaiPhong_Load(sender, e);
+                    DialogResult kq = MessageBox.Show("Cập nhật thành công!", "Thông báo");
+                   if (kq == DialogResult.Yes)
+                    {
+                        frmLoaiPhong_Load(sender, e);
+                    }
                 }
